nfc: working transparent mode

+++ lib/nfc/protocols/mf_classic/crypto1.c
@@ -127,7 +127,7 @@ void crypto1_encrypt(Crypto1* crypto, uint8_t* keystream, const BitBuffer* buff,
         }
         bit_buffer_set_byte(out, 0, encrypted_byte);
     } else {
-        for(uint8_t i = 0; i < bits / 8; i++) {
+        for(size_t i = 0; i < bits / 8; i++) {
             uint8_t encrypted_byte = crypto1_byte(crypto, keystream ? keystream[i] : 0, 0) ^
                                      plain_data[i];
             bool parity_bit =
